requirements.txt
fastapi
gunicorn
uvicorn
python-multipart
tensorflow
numpy
pillow
requests
pydantic